+++ sx1262_lora_gateway/run.sh
@@ -8,6 +8,9 @@ LORA_SF=$(bashio::config 'lora_spreading_factor')
 LORA_BW=$(bashio::config 'lora_bandwidth')
 LORA_CR=$(bashio::config 'lora_coding_rate')
 LORA_SW=$(bashio::config 'lora_sync_word')
+LORA_SW_FORCE=$(bashio::config 'lora_sync_word_force')
+LORA_SW_MSB=$(bashio::config 'lora_sync_word_msb')
+LORA_SW_LSB=$(bashio::config 'lora_sync_word_lsb')
 LORA_POWER=$(bashio::config 'lora_tx_power')
 MQTT_HOST=$(bashio::config 'mqtt_host')
 MQTT_PORT=$(bashio::config 'mqtt_port')
@@ -22,7 +25,7 @@ bashio::log.info "Bandwidth: ${LORA_BW} Hz"
 bashio::log.info "MQTT Broker: ${MQTT_HOST}:${MQTT_PORT}"
 
 # Export config as environment variables
-export LORA_FREQ LORA_SF LORA_BW LORA_CR LORA_SW LORA_POWER
+export LORA_FREQ LORA_SF LORA_BW LORA_CR LORA_SW LORA_SW_FORCE LORA_SW_MSB LORA_SW_LSB LORA_POWER
 export MQTT_HOST MQTT_PORT MQTT_USER MQTT_PASS MQTT_PREFIX
 export LOG_LEVEL